update for next release
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/CNORfuzzy@79889 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/cSimulator.R
@@ -53,9 +53,9 @@ cSimulator <- function(CNOlist, model, simList) {
 	indexInhibitors <- as.integer((indexList$inhibited)-1)
     nSignals <- length(indexSignals)
 
-	# cnolist
-	valueInhibitors <- as.integer(CNOlist@inhibitors)
-	valueStimuli <- as.integer(CNOlist@stimuli)
+	# cnolist changed to float in bug report 222
+	valueInhibitors <- as.numeric(CNOlist@inhibitors)
+	valueStimuli <- as.numeric(CNOlist@stimuli)
 
 
 	res = .Call("simulatorT1",
